Make controls internal and main exe able to see the internals of the core dll.
